mcp/server/requirements.txt
fastmcp
psycopg2-binary
requests
beautifulsoup4
python-dotenv
numpy
pydantic
spacy
sentence-transformers
google-generativeai